requirements.txt
Django==2.0 
djangorestframework==3.9.1 
